...code code code...

+++ osrm/update_osrm.sls
@@ -27,9 +27,9 @@ osrm_reindex_{{instance.profile}}:
         ./osrm-extract extract.osm.pbf # creates .names, ., .restrictions
         ./osrm-prepare extract.osrm # creates .fileIndex, .hsgr, .nodes, .ramIndex, .edges
         sleep 2
-        #test -f extract.osrm.hsgr || ( echo "OSRM extract failed somehow." && exit 1 )
+        test -f extract.osrm.hsgr || ( echo "OSRM extract failed somehow." && exit 1 )
         pgrep pkill -f 'osrm-routed.*-p {{ instance.port }}' || echo "Started osrm-routed." # Make sure we kill the right instance.
         {{ pillar.tm_dir}}/log.sh "OSRM index for profile '{{ instance.name}}' rebuilt."
-        exit 0 # so Salt doesn't report failed pkill as a fail.
+        
     - watch: [ cmd: osrm_start_{{instance.profile}} ]
 {% endfor %}
